fix: use filterset
