+++ pkg/query-service/app/integrations/messagingQueues/celery/translator.go
@@ -1,0 +1,1 @@
+package celery
