+++ pkg/query-service/app/integrations/messagingQueues/kafka/model.go
@@ -22,37 +22,3 @@ type OnboardingResponse struct {
 	Message   string `json:"error_message"`
 	Status    string `json:"status"`
 }
-
-// QueueFilters
-// ToDo: add capability of dynamic filtering based on any of the filters
-type QueueFilters struct {
-	ServiceName []string
-	SpanName    []string
-	Queue       []string
-	Destination []string
-	Kind        []string
-}
-
-type CeleryTask struct {
-	kind   string
-	status string
-}
-
-type CeleryTasks interface {
-	GetKind() string
-	GetStatus() string
-	Set(string, string)
-}
-
-func (r *CeleryTask) GetKind() string {
-	return r.kind
-}
-
-func (r *CeleryTask) GetStatus() string {
-	return r.status
-}
-
-func (r *CeleryTask) Set(kind, status string) {
-	r.kind = kind
-	r.status = status
-}
