+++ pkg/query-service/app/integrations/messagingQueues/kafka/translator.go
@@ -2,11 +2,9 @@ package kafka
 
 import (
 	"fmt"
-
 	"go.signoz.io/signoz/pkg/query-service/common"
 	"go.signoz.io/signoz/pkg/query-service/constants"
 	v3 "go.signoz.io/signoz/pkg/query-service/model/v3"
-	"strings"
 )
 
 var defaultStepInterval int64 = 60
@@ -21,6 +19,7 @@ func BuildQueryRangeParams(messagingQueue *MessagingQueue, queryContext string) 
 	queueType := KafkaQueue
 
 	chq, err := BuildClickHouseQuery(messagingQueue, queueType, queryContext)
+
 	if err != nil {
 		return nil, err
 	}
@@ -80,10 +79,9 @@ func buildBuilderQueriesProducerBytes(
 			Type:     v3.AttributeKeyType("Gauge"),
 			IsColumn: true,
 		},
-		AggregateOperator: v3.AggregateOperatorAvg,
-		Temporality:       v3.Unspecified,
-		TimeAggregation:   v3.TimeAggregationAvg,
-		SpaceAggregation:  v3.SpaceAggregationAvg,
+		Temporality:      v3.Unspecified,
+		TimeAggregation:  v3.TimeAggregationAvg,
+		SpaceAggregation: v3.SpaceAggregationAvg,
 		Filters: &v3.FilterSet{
 			Operator: "AND",
 			Items: []v3.FilterItem{
@@ -321,34 +319,6 @@ func BuildQRParamsWithCache(
 	return queryRangeParams, err
 }
 
-func getFilters(variables map[string]string) *QueueFilters {
-	return &QueueFilters{
-		ServiceName: parseFilter(variables["service_name"]),
-		SpanName:    parseFilter(variables["span_name"]),
-		Queue:       parseFilter(variables["queue"]),
-		Destination: parseFilter(variables["destination"]),
-		Kind:        parseFilter(variables["kind"]),
-	}
-}
-
-// parseFilter splits a comma-separated string into a []string.
-// Returns an empty slice if the input is blank.
-func parseFilter(val string) []string {
-	if val == "" {
-		return []string{}
-	}
-	// Split on commas, trim whitespace around each part
-	parts := strings.Split(val, ",")
-	var out []string
-	for _, p := range parts {
-		trimmed := strings.TrimSpace(p)
-		if trimmed != "" {
-			out = append(out, trimmed)
-		}
-	}
-	return out
-}
-
 func BuildClickHouseQuery(
 	messagingQueue *MessagingQueue,
 	queueType string,
@@ -385,8 +355,8 @@ func BuildClickHouseQuery(
 	var query string
 
 	switch queryContext {
-	case "overview":
-		query = generateOverviewSQL(start, end, getFilters(messagingQueue.Variables))
+	//case "overview":
+	//	query = generateOverviewSQL(start, end, getFilters(messagingQueue.Variables))
 	case "producer":
 		query = generateProducerSQL(start, end, topic, partition, queueType)
 	case "consumer":
